Move arguments into a separate file
